seperated test types into directories
