+++ tests/regression/test_classification_pipeline.py
