+++ tests/regression/test_detection_pipeline.py
