+++ tests/regression/test_extraction_pipeline.py
