+++ tests/smoke/test_cli.py
