+++ tests/smoke/test_import.py
